Use decimal field instead of float field

+++ bluebottle/bb_tasks/serializers.py
@@ -54,7 +54,7 @@ class BaseTaskSerializer(TaggableSerializerMixin, serializers.ModelSerializer):
     status = serializers.ChoiceField(choices=BB_TASK_MODEL.TaskStatuses.choices,
                                      default=BB_TASK_MODEL.TaskStatuses.open)
     tags = TagSerializer()
-    time_needed = serializers.FloatField(min_value=0.0)
+    time_needed = serializers.DecimalField(min_value=0.0)
 
     meta_data = MetaField(
         title='get_meta_title',
